grouped token elements in group elemet
grouped the elements in the html for the tokens, then adjusted the script to still function with the group. Instead of adjusting the 'cx' and 'cy' value of the element, adjusts the transform of the group. This should allow for a multiple objected border, and allows for greater flexibility later on