updated organization
updated the organization of the token information & navbar for home page and board page. Removed the a tags and cleaned up the css as well for buttons. changed naming scheme as well to correctly label buttons and non button elements. Also added a bottom border to the token information element